+++ src/main/java/pro/i_it/learn/server/module/interfaces/ISocketIOManager.java
@@ -7,5 +7,4 @@ import com.corundumstudio.socketio.SocketIOServer;
  */
 public interface ISocketIOManager {
     ISender getSocketIOSender();
-    SocketIOServer getServer();
 }
