+++ src/main/java/pro/i_it/learn/server/module/socketIOManager/pansmile/SocketIOManager.java
@@ -20,27 +20,6 @@ public class SocketIOManager implements ISocketIOManager {
     }
 
 
-    private void initServer() {
-        if (server == null) {
-            config = new Configuration();
-            config.setPort(8080);
-            config.setHostname("localhost");
-            server = new SocketIOServer(config);
-            System.out.println("SERVER CREATED");
-            server.addEventListener("MESSAGE", String.class,
-                    (client, message, ackRequest) -> {
-                        System.out.println("Client said: " + message);
-                    });
-
-        }
-    }
-    public SocketIOServer getServer() {
-        if (server == null) {
-            initServer();
-        }
-        return server;
-    }
-
     @Override
     public ISender getSocketIOSender() {
